Modularized settings into it's own view file
